hal: fix SD card HAL initialization when no SD card is inserted (#12)
Rev 1.0 of the flight controller has the SD card detect switch incorrectly hardwired to ground. So the SD card detect pin always reports an SD card as being inserted.

Fix is to use a macro to remove the call to the error handler, so the flight controller does not lock up when an SD card is not inserted.